Breakout dashboard blueprint into a new layout.
